requirements.txt
Django>=3.0,<4.0
psycopg2>=2.7,<3.0
djangorestframework==3.11.0
ipython
django-extensions==2.1.6
black